backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clerk-backend-api>=3.1.11",
    "fastapi>=0.116.1",
    "openai>=1.99.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "svix>=1.74.1",
    "uvicorn>=0.35.0",
]
